# Conflicts:
#	ART/android/sub_app/fermat-art-android-sub-app-artist-community-bitdubai/src/main/java/com/bitdubai/sub_app_artist_community/fragments/ConnectionsListFragment.java
#	ART/android/sub_app/fermat-art-android-sub-app-artist-community-bitdubai/src/main/java/com/bitdubai/sub_app_artist_community/fragments/ListUserIdentiesFragment.java
#	ART/android/sub_app/fermat-art-android-sub-app-artist-identity-bitdubai/src/main/java/com/bitdubai/fermat_art_android_sub_app_artist_identity_bitdubai/fragments/CreateArtistIndetityFragment.java
#	ART/android/sub_app/fermat-art-android-sub-app-artist-identity-bitdubai/src/main/java/com/bitdubai/fermat_art_android_sub_app_artist_identity_bitdubai/popups/PresentationArtArtistUserIdentityDialog.java
#	ART/android/sub_app/fermat-art-android-sub-app-fan-community-bitdubai/src/main/java/com/bitdubai/sub_app/fan_community/notifications/CommunityNotificationPainterBuilder.java
#	ART/android/sub_app/fermat-art-android-sub-app-fan-identity-bitdubai/src/main/java/com/bitdubai/sub_app/art_fan_identity/fragment_factory/ArtFanUserIdentityFragmentFactory.java
#	ART/android/sub_app/fermat-art-android-sub-app-fan-identity-bitdubai/src/main/java/com/bitdubai/sub_app/art_fan_identity/fragments/CreateArtFanUserIdentityFragment.java
#	ART/android/sub_app/fermat-art-android-sub-app-fan-identity-bitdubai/src/main/java/com/bitdubai/sub_app/art_fan_identity/popup/PresentationArtFanUserIdentityDialog.java
#	ART/library/api/fermat-art-api/build.gradle
#	ART/plugin/actor_network_service/fermat-art-plugin-actor-network-service-artist-bitdubai/src/main/java/com/bitdubai/fermat_art_plugin/layer/actor_network_service/artist/developer/bitdubai/version_1/database/ArtistActorNetworkServiceDao.java
#	ART/plugin/actor_network_service/fermat-art-plugin-actor-network-service-fan-bitdubai/src/main/java/com/bitdubai/fermat_art_plugin/layer/actor_network_service/fan/developer/bitdubai/version_1/messages/NetworkServiceMessage.java
#	ART/plugin/actor_network_service/fermat-art-plugin-actor-network-service-fan-bitdubai/src/main/java/com/bitdubai/fermat_art_plugin/layer/actor_network_service/fan/developer/bitdubai/version_1/messages/RequestMessage.java
#	ART/plugin/actor_network_service/fermat-art-plugin-actor-network-service-fan-bitdubai/src/main/java/com/bitdubai/fermat_art_plugin/layer/actor_network_service/fan/developer/bitdubai/version_1/structure/FanActorNetworkServiceSearch.java
#	CHT/android/sub_app/fermat-cht-android-sub-app-chat-identity-bitdubai/src/main/java/com/bitbudai/fermat_cht_android_sub_app_chat_identity_bitdubai/util/DialogCropImage.java
#	DMP/android/sub_app/fermat-dmp-android-sub-app-wallet-manager-bitdubai/src/main/java/com/bitdubai/sub_app/wallet_manager/fragment/CommunitiesExpandibleFragment.java
#	TKY/android/reference_wallet/fermat-tky-android-reference-wallet-fan-wallet-bitdubai/src/main/java/com/bitdubai/reference_wallet/fan_wallet/session/FanWalletSession.java
#	TKY/android/sub_app/fermat-tky-android-sub-app-artist-identity-bitdubai/src/main/java/com/bitdubai/fermat_tky_android_sub_app_artist_identity_bitdubai/fragments/TokenlyArtistIdentityCreateProfile.java
#	TKY/android/sub_app/fermat-tky-android-sub-app-artist-identity-bitdubai/src/main/java/com/bitdubai/fermat_tky_android_sub_app_artist_identity_bitdubai/popup/PresentationTokenlyArtistUserIdentityDialog.java
#	TKY/android/sub_app/fermat-tky-android-sub-app-fan-identity-bitdubai/src/main/java/com/bitdubai/sup_app/tokenly_fan_user_identity/fragments/CreateTokenlyFanUserIdentityFragment.java
#	TKY/android/sub_app/fermat-tky-android-sub-app-fan-identity-bitdubai/src/main/java/com/bitdubai/sup_app/tokenly_fan_user_identity/popup/PresentationTokenlyFanUserIdentityDialog.java
#	WPD/android/sub_app/fermat-wpd-android-sub-app-wallet-manager-bitdubai/src/main/res/drawable/artist_banner.jpg
#	WPD/plugin/engine/fermat-wpd-plugin-engine-app-runtime-bitdubai/src/main/java/com/bitdubai/fermat_wpd_plugin/layer/engine/wallet_runtime/developer/bitdubai/version_1/AppRuntimeEnginePluginRoot.java
#	WPD/plugin/engine/fermat-wpd-plugin-engine-app-runtime-bitdubai/src/main/java/com/bitdubai/fermat_wpd_plugin/layer/engine/wallet_runtime/developer/bitdubai/version_1/provisory/SubAppAppsGenerator.java

+++ BNK/library/api/fermat-bnk-api/src/main/java/com/bitdubai/fermat_bnk_api/layer/bnk_wallet/bank_money/interfaces/BankMoneyWalletBalance.java
@@ -3,6 +3,8 @@ package com.bitdubai.fermat_bnk_api.layer.bnk_wallet.bank_money.interfaces;
 import com.bitdubai.fermat_bnk_api.all_definition.enums.BalanceType;
 import com.bitdubai.fermat_bnk_api.layer.bnk_wallet.bank_money.exceptions.*;
 
+import java.math.BigDecimal;
+
 
 /**
  * Created by Yordin Alayn on 30.09.15.
@@ -11,7 +13,7 @@ import com.bitdubai.fermat_bnk_api.layer.bnk_wallet.bank_money.exceptions.*;
 public interface BankMoneyWalletBalance {
 
 
-    double getBalance(String accountNumber)  throws CantCalculateBalanceException;
+    BigDecimal getBalance(String accountNumber)  throws CantCalculateBalanceException;
 
     void debit(BankMoneyTransactionRecord bankMoneyTransactionRecord) throws CantRegisterDebitException;
