+++ BNK/plugin/bank_money_transaction/fermat-bnk-plugin-bank-money-transaction-hold-bitdubai/src/main/java/com/bitdubai/fermat_bnk_plugin/layer/bank_money_transaction/hold/developer/bitdubai/version_1/structure/BankTransactionImpl.java
@@ -7,6 +7,7 @@ import com.bitdubai.fermat_bnk_api.all_definition.enums.BankTransactionStatus;
 import com.bitdubai.fermat_bnk_api.all_definition.enums.TransactionType;
 
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
@@ -23,7 +24,7 @@ public class BankTransactionImpl implements BankTransaction {
 
     private String publicKeyActor;
 
-    private float amount;
+    private BigDecimal amount;
 
     private String accountNumber;
 
@@ -39,7 +40,7 @@ public class BankTransactionImpl implements BankTransaction {
 
     private BankTransactionStatus status;
 
-    public BankTransactionImpl(UUID transactionId, String publicKeyPlugin, String publicKeyWallet, String publicKeyActor, float amount, String accountNumber, FiatCurrency fiatCurrency, String memo, BankOperationType bankOperationType, TransactionType transactionType, long timestamp,BankTransactionStatus status) {
+    public BankTransactionImpl(UUID transactionId, String publicKeyPlugin, String publicKeyWallet, String publicKeyActor, BigDecimal amount, String accountNumber, FiatCurrency fiatCurrency, String memo, BankOperationType bankOperationType, TransactionType transactionType, long timestamp,BankTransactionStatus status) {
         this.transactionId = transactionId;
         this.publicKeyPlugin = publicKeyPlugin;
         this.publicKeyWallet = publicKeyWallet;
@@ -75,7 +76,7 @@ public class BankTransactionImpl implements BankTransaction {
     }
 
     @Override
-    public float getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
