+++ BNK/plugin/wallet/fermat-bnk-plugin-wallet-bank-money-bitdubai/src/main/java/com/bitdubai/fermat_bnk_plugin/layer/wallet/bank_money/developer/bitdubai/version_1/structure/BankMoneyTransactionRecordImpl.java
@@ -1,5 +1,6 @@
 package com.bitdubai.fermat_bnk_plugin.layer.wallet.bank_money.developer.bitdubai.version_1.structure;
 
+import com.bitdubai.fermat_api.layer.all_definition.common.system.interfaces.error_manager.enums.UnexpectedPluginExceptionSeverity;
 import com.bitdubai.fermat_api.layer.all_definition.enums.FiatCurrency;
 import com.bitdubai.fermat_bnk_api.all_definition.enums.BalanceType;
 import com.bitdubai.fermat_bnk_api.all_definition.enums.BankAccountType;
@@ -8,8 +9,10 @@ import com.bitdubai.fermat_bnk_api.all_definition.enums.BankTransactionStatus;
 import com.bitdubai.fermat_bnk_api.all_definition.enums.TransactionType;
 import com.bitdubai.fermat_bnk_api.layer.bnk_wallet.bank_money.interfaces.BankMoneyTransactionRecord;
 
+import java.math.BigDecimal;
 import java.io.Serializable;
 import java.util.UUID;
+
 
 /**
  * Created by francisco on 05/11/15.
@@ -19,20 +22,20 @@ public class BankMoneyTransactionRecordImpl implements BankMoneyTransactionRecor
     UUID bankTransactionId;
     String balanceType;
     String transactionType;
-    float amount;
+    BigDecimal amount;
     String cashCurrencyType;
     String bankOperationType;
     String bankDocumentReference;
     String bankName;
     String bankAccountNumber;
     String bankAccountType;
-    long runningBookBalance;
-    long runningAvailableBalance;
+    BigDecimal runningBookBalance;
+    BigDecimal runningAvailableBalance;
     long timeStamp;
     String memo;
     String status;
 
-    public BankMoneyTransactionRecordImpl(UUID bankTransactionId, String balanceType, String transactionType, float amount, String cashCurrencyType, String bankOperationType, String bankDocumentReference, String bankName, String bankAccountNumber, String bankAccountType, long runningBookBalance, long runningAvailableBalance, long timeStamp, String memo, String status) {
+    public BankMoneyTransactionRecordImpl(UUID bankTransactionId, String balanceType, String transactionType, BigDecimal amount, String cashCurrencyType, String bankOperationType, String bankDocumentReference, String bankName, String bankAccountNumber, String bankAccountType, BigDecimal runningBookBalance, BigDecimal runningAvailableBalance, long timeStamp, String memo, String status) {
         this.bankTransactionId = bankTransactionId;
         this.balanceType = balanceType;
         this.transactionType = transactionType;
@@ -67,7 +70,7 @@ public class BankMoneyTransactionRecordImpl implements BankMoneyTransactionRecor
 
     @Override
     public TransactionType getTransactionType() {
-        try{
+        try {
             return TransactionType.getByCode(transactionType);
         }catch (Exception e){
             return null;
@@ -75,7 +78,7 @@ public class BankMoneyTransactionRecordImpl implements BankMoneyTransactionRecor
     }
 
     @Override
-    public float getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
@@ -90,7 +93,7 @@ public class BankMoneyTransactionRecordImpl implements BankMoneyTransactionRecor
 
     @Override
     public BankOperationType getBankOperationType() {
-        try{
+        try {
             return BankOperationType.getByCode(bankOperationType);
         }catch (Exception e){
             return null;
@@ -127,12 +130,12 @@ public class BankMoneyTransactionRecordImpl implements BankMoneyTransactionRecor
     }
 
     @Override
-    public long getRunningBookBalance() {
+    public BigDecimal getRunningBookBalance() {
         return runningBookBalance;
     }
 
     @Override
-    public long getRunningAvailableBalance() {
+    public BigDecimal getRunningAvailableBalance() {
         return runningAvailableBalance;
     }
